make design change, change algorithm of transforming Image to RGBImage and back

+++ JPEGAlgorithm/Averager.cs
@@ -13,7 +13,7 @@ namespace JPEGAlgorithm {
                 return imageBlock;
             }
             var array = imageBlock.Data;
-            double average;
+            float average;
             for (var i = 0; i < w; i+=2) {
                 for (var j = 0; j < h; j+=2) {
                     average = (array[i, j] + array[i+1, j] + array[i, j+1] + array[i+1, j+1])/4;
@@ -29,13 +29,13 @@ namespace JPEGAlgorithm {
             }
             int[,] data;
             int[,] resData = new int[8, 8];
-            double average;
+            float average;
             for (var k = 0; k < 2; k++) {
                 for (var t = 0; t < 2; t++) {
                     data = imageBlocksMatrix[k, t].Data;
                     for (int i = 4 * k, u = 0; i < 4 * (1 + k); i++, u+=2) {
                         for (int j = 4 * t, v = 0; j < 4 * (1 + t); j++, v+=2) {
-                            average = ((double)(data[u, v] + data[u + 1, v] + data[u, v + 1] + data[u + 1, v + 1])) / 4;
+                            average = ((float)(data[u, v] + data[u + 1, v] + data[u, v + 1] + data[u + 1, v + 1])) / 4;
                             resData[i, j] = (int)Math.Round(average);
                         }
                     }
